+++ meta-facebook/yv35-bb/src/platform/plat_def.h
@@ -1,4 +1,6 @@
 #ifndef PLAT_DEF_H
 #define PLAT_DEF_H
 
+#define BMC_USB_PORT "CDC_ACM_0"
+
 #endif
