+++ meta-facebook/yv35-bb/src/platform/plat_version.h
@@ -1,6 +1,8 @@
 #ifndef PLAT_VERSION_H
 #define PLAT_VERSION_H
 
+#define PLATFORM_NAME "yv35"
+#define PROJECT_NAME "baseboard"
 #define IANA_ID 0x009c9c
 #define DEVICE_ID 0x00
 #define DEVICE_REVISION 0x80
