+++ meta-facebook/yv35-cl/CMakeLists.txt
@@ -6,12 +6,11 @@ find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(yv35-cl)
 
 set(common_path ../../common)
-
-FILE(GLOB app_sources src/*.c src/ipmi/*.c src/sensor/*.c src/sensor/dev/*.c src/platform/*.c src/lib/*.c)
+FILE(GLOB app_sources src/ipmi/*.c src/platform/*.c src/lib/*.c)
 FILE(GLOB common_sources ${common_path}/*.c ${common_path}/ipmi/*.c ${common_path}/host/*.c ${common_path}/sensor/*.c ${common_path}/usb/*.c ${common_path}/hal/*.c ${common_path}/dev/*.c ${common_path}/lib/*.c ${common_path}/ipmb/*.c)
 
 target_sources(app PRIVATE ${app_sources})
 target_sources(app PRIVATE ${common_sources})
 target_include_directories(app PRIVATE ${ZEPHYR_BASE}/include/portability)
 target_include_directories(app PRIVATE ${common_path} ${common_path}/ipmi/include ${common_path}/host ${common_path}/sensor ${common_path}/hal ${common_path}/dev ${common_path}/lib ${common_path}/usb ${common_path}/ipmb)
-target_include_directories(app PRIVATE src/ipmi/include src/lib src/platform src/sensor)
+target_include_directories(app PRIVATE src/ipmi/include src/lib src/platform)
