+++ meta-facebook/yv35-cl/src/platform/plat_def.h
@@ -4,4 +4,7 @@
 #define ENABLE_ASD
 #define ENABLE_ISL69260
 
+#define HOST_KCS_PORT kcs3
+#define BMC_USB_PORT "CDC_ACM_0"
+
 #endif
